// src/modules/product/models/responses/ProductResponse.ts
import {BaseAuditResponse} from 'common';
import {CollectionResponse} from './CollectionResponse';
import {VariantResponse} from './VariantResponse';
import {AttributeResponse} from './AttributeResponse';

export interface ProductResponse extends BaseAuditResponse {
  id: number;
  advantages: string;
  available: number | null;
  brand: string;
  brandName: string;
  careLabels: string;
  category: string | null;
  categoryId: number | null;
  code: string;
  collections: Array<CollectionResponse>;
  component: string | null;
  content: null;
  defect: string | null;
  description: string | null;
  designer: string | null;
  designerCode: string | null;
  goods: string;
  goodsName: string;
  madeIn: string;
  madeInId: number;
  material: string | null;
  materialAdvantages: string | null;
  materialApplication: string | null;
  materialComponent: string | null;
  materialDefect: string | null;
  materialId: number | null;
  merchandiser: string | null;
  merchandiserCode: string | null;
  name: string;
  onHand: number | null;
  productType: string;
  referenceId: number | null;
  specifications: string | null;
  status: string | null;
  tags: string;
  unit: string | null;
  unitName: string | null;
  variants: Array<VariantResponse>;
  attributes: Array<AttributeResponse>;
}